Remove unnecessary interfaces, cleanup code

+++ Task1/Frequency.cs
@@ -3,7 +3,7 @@ namespace Task1
 {
     public record Frequency
     {
-        public float Percent { get; init; }
+        public double Percent { get; init; }
 
         public override string ToString()
         {
